backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
anthropic==0.42.0
openai==1.57.4
python-multipart==0.0.17
aiofiles==24.1.0
pydantic==2.10.3
python-dotenv==1.0.1
reportlab==4.2.5
pypdf==5.1.0
python-docx==1.1.2
Pillow>=10.0.0,<11
pytesseract==0.3.13
